Major restructuring. Work first, fix later.
